Revert bits that aren't specifically to do with loading DRM files
